requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.6
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.17.3
packaging==23.1
Pillow==9.5.0
pluggy==1.2.0
PyJWT==2.7.0
pyrsistent==0.19.3
pytest==7.4.0
pytest-django==4.5.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
sqlparse==0.4.4
swagger-spec-validator==3.0.3
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==2.0.3
faker